+++ example/lib/main.dart
@@ -12,73 +12,74 @@ class MyApp extends StatelessWidget {
         primarySwatch: Colors.blue,
       ),
       debugShowCheckedModeBanner: false,
-      home: MyHomePage(title: 'Filter list plugin demo'),
+      home: MyHomePage(title: 'Filter list example'),
     );
   }
 }
 
 class MyHomePage extends StatefulWidget {
-  MyHomePage({Key key, this.title}) : super(key: key);
-  final String title;
+  MyHomePage({Key? key, this.title}) : super(key: key);
+  final String? title;
 
   @override
   _MyHomePageState createState() => _MyHomePageState();
 }
 
 class _MyHomePageState extends State<MyHomePage> {
-  List<User> selectedUserList = [];
+  List<User>? selectedUserList = [];
 
   void _openFilterDialog() async {
-    await FilterListDialog.display(
+    await FilterListDialog.display<User>(
       context,
       listData: userList,
       selectedListData: selectedUserList,
       height: 480,
       headlineText: "Select Users",
       searchFieldHintText: "Search Here",
-      label: (item) {
-        return item.name;
+      choiceChipLabel: (item) {
+        return item!.name;
       },
       validateSelectedItem: (list, val) {
-        return list.contains(val);
+        return list!.contains(val);
       },
 
       onItemSearch: (list, text) {
-        if (list.any((element) =>
-            element.name.toLowerCase().contains(text.toLowerCase()))) {
-          /// return list which contains matches
-          return list
-              .where((element) =>
-                  element.name.toLowerCase().contains(text.toLowerCase()))
-              .toList();
+        if (list != null) {
+          if (list.any((element) =>
+              element.name!.toLowerCase().contains(text.toLowerCase()))) {
+            /// return list which contains matches
+            return list
+                .where((element) =>
+                    element.name!.toLowerCase().contains(text.toLowerCase()))
+                .toList();
+          }
         }
-        return null;
+
+        return [];
       },
 
       onApplyButtonClick: (list) {
-        if (list != null) {
-          setState(() {
-            selectedUserList = List.from(list);
-          });
-          Navigator.pop(context);
-        }
+        setState(() {
+          selectedUserList = List.from(list!);
+        });
+        Navigator.pop(context);
       },
 
       /// uncomment below code to create custom choice chip
-      /*  choiceChipBuilder: (context, item, isSelected) {
-        return Container(
-          padding: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
-          margin: EdgeInsets.symmetric(horizontal: 10, vertical: 10),
-          decoration: BoxDecoration(
-              border: Border.all(
-            color: isSelected ? Colors.blue[300] : Colors.grey[300],
-          )),
-          child: Text(
-            item.name,
-            style: TextStyle(
-                color: isSelected ? Colors.blue[300] : Colors.grey[300]),
-          ),
-        );
+      /* choiceChipBuilder: (context, item, isSelected) {
+          return Container(
+            padding: EdgeInsets.symmetric(horizontal: 16, vertical: 12),
+            margin: EdgeInsets.symmetric(horizontal: 10, vertical: 10),
+            decoration: BoxDecoration(
+                border: Border.all(
+              color: isSelected! ? Colors.blue[300]! : Colors.grey[300]!,
+            )),
+            child: Text(
+              item.name,
+              style: TextStyle(
+                  color: isSelected ? Colors.blue[300] : Colors.grey[300]),
+            ),
+          );
       },*/
     );
   }
@@ -87,7 +88,7 @@ class _MyHomePageState extends State<MyHomePage> {
   Widget build(BuildContext context) {
     return Scaffold(
       appBar: AppBar(
-        title: Text(widget.title),
+        title: Text(widget.title!),
       ),
       bottomNavigationBar: Row(
         mainAxisAlignment: MainAxisAlignment.spaceEvenly,
@@ -130,7 +131,7 @@ class _MyHomePageState extends State<MyHomePage> {
       ),
       body: Column(
         children: <Widget>[
-          selectedUserList == null || selectedUserList.length == 0
+          selectedUserList == null || selectedUserList!.length == 0
               ? Expanded(
                   child: Center(
                     child: Text('No text selected'),
@@ -140,11 +141,11 @@ class _MyHomePageState extends State<MyHomePage> {
                   child: ListView.separated(
                       itemBuilder: (context, index) {
                         return ListTile(
-                          title: Text(selectedUserList[index].name),
+                          title: Text(selectedUserList![index].name!),
                         );
                       },
                       separatorBuilder: (context, index) => Divider(),
-                      itemCount: selectedUserList.length),
+                      itemCount: selectedUserList!.length),
                 ),
         ],
       ),
@@ -153,10 +154,10 @@ class _MyHomePageState extends State<MyHomePage> {
 }
 
 class FilterPage extends StatelessWidget {
-  const FilterPage({Key key, this.allTextList, this.selectedUserList})
+  const FilterPage({Key? key, this.allTextList, this.selectedUserList})
       : super(key: key);
-  final List<User> allTextList;
-  final List<User> selectedUserList;
+  final List<User>? allTextList;
+  final List<User>? selectedUserList;
   @override
   Widget build(BuildContext context) {
     return Scaffold(
@@ -164,16 +165,16 @@ class FilterPage extends StatelessWidget {
         title: Text("Filter list Page"),
       ),
       body: SafeArea(
-        child: FilterListWidget(
+        child: FilterListWidget<User>(
           listData: userList,
           selectedListData: selectedUserList,
-          hideheaderText: true,
+          hideHeaderText: true,
           onApplyButtonClick: (list) {
             Navigator.pop(context, list);
           },
-          label: (item) {
+          choiceChipLabel: (item) {
             /// Used to print text on chip
-            return item.name;
+            return item!.name;
           },
           choiceChipBuilder: (context, item, isSelected) {
             return Container(
@@ -181,28 +182,28 @@ class FilterPage extends StatelessWidget {
               margin: EdgeInsets.symmetric(horizontal: 10, vertical: 10),
               decoration: BoxDecoration(
                   border: Border.all(
-                color: isSelected ? Colors.blue[300] : Colors.grey[300],
+                color: isSelected! ? Colors.blue[300]! : Colors.grey[300]!,
               )),
               child: Text(item.name),
             );
           },
           validateSelectedItem: (list, val) {
             ///  identify if item is selected or not
-            return list.contains(val);
+            return list!.contains(val);
           },
           onItemSearch: (list, text) {
             /// When text change in search text field then return list containing that text value
             ///
             ///Check if list has value which matchs to text
-            if (list.any((element) =>
-                element.name.toLowerCase().contains(text.toLowerCase()))) {
+            if (list!.any((element) =>
+                element.name!.toLowerCase().contains(text.toLowerCase()))) {
               /// return list which contains matches
               return list
                   .where((element) =>
-                      element.name.toLowerCase().contains(text.toLowerCase()))
+                      element.name!.toLowerCase().contains(text.toLowerCase()))
                   .toList();
             }
-            return null;
+            return [];
           },
         ),
       ),
@@ -211,50 +212,50 @@ class FilterPage extends StatelessWidget {
 }
 
 class User {
-  final String name;
-  final String avatar;
+  final String? name;
+  final String? avatar;
   User({this.name, this.avatar});
 }
 
 /// Creating a global list for example purpose.
-/// Generally it should be within data class or whereever you want
+/// Generally it should be within data class or where ever you want
 List<User> userList = [
-  User(name: "Jon", avatar: "asd"),
-  User(name: "Lindsey ", avatar: "asd"),
-  User(name: "Valarie ", avatar: "asd"),
-  User(name: "Elyse ", avatar: "asd"),
-  User(name: "Ethel ", avatar: "asd"),
-  User(name: "Emelyan ", avatar: "asd"),
-  User(name: "Catherine ", avatar: "asd"),
-  User(name: "Stepanida  ", avatar: "asd"),
-  User(name: "Carolina ", avatar: "asd"),
-  User(name: "Nail  ", avatar: "asd"),
+  User(name: "Jon", avatar: "user.png"),
+  User(name: "Lindsey ", avatar: "user.png"),
+  User(name: "Valarie ", avatar: "user.png"),
+  User(name: "Elyse ", avatar: "user.png"),
+  User(name: "Ethel ", avatar: "user.png"),
+  User(name: "Emelyan ", avatar: "user.png"),
+  User(name: "Catherine ", avatar: "user.png"),
+  User(name: "Stepanida  ", avatar: "user.png"),
+  User(name: "Carolina ", avatar: "user.png"),
+  User(name: "Nail  ", avatar: "user.png"),
 ];
 
 /// Another exmaple of [FilterListWidget] to filter list of strings
 ///
-/// FilterListWidget(
+/// FilterListWidget<String>(
 ///   listData: ["One", "Two", "Three", "Four","five","Six","Seven","Eight","Nine","Ten"],
 ///   selectedListData: ["One", "Three", "Four","Eight","Nine"],
-///   hideheaderText: true,
+///   hideHeaderText: true,
 ///   height: MediaQuery.of(context).size.height,
-///   // hideheaderText: true,
+///   // hideHeaderText: true,
 ///   onApplyButtonClick: (list) {
 ///     Navigator.pop(context, list);
 ///   },
-///   label: (item) {
+///   choiceChipLabel: (item) {
 ///     /// Used to print text on chip
 ///     return item;
 ///   },
 ///   validateSelectedItem: (list, val) {
 ///     ///  identify if item is selected or not
-///     return list.contains(val);
+///     return list!.contains(val);
 ///   },
 ///   onItemSearch: (list, text) {
 ///     /// When text change in search text field then return list containing that text value
 ///     ///
 ///     ///Check if list has value which matchs to text
-///     if (list.any((element) =>
+///     if (list!.any((element) =>
 ///         element.toLowerCase().contains(text.toLowerCase()))) {
 ///       /// return list which contains matches
 ///       return list
@@ -262,5 +263,6 @@ List<User> userList = [
 ///               element.toLowerCase().contains(text.toLowerCase()))
 ///           .toList();
 ///     }
+///     return [];
 ///   },
 /// )