+++ lib/filter_list.dart
@@ -1,1 +1,3 @@
+library filter_list;
+
 export 'src/filter_list_dialog.dart';