+++ lib/src/filter_list_dialog.dart
@@ -7,132 +7,185 @@ import 'package:flutter/rendering.dart';
 
 part 'filter_list_widget.dart';
 
+/// The [FilterListDialog.display()] is a [Dialog] with some filter utilities and callbacks which helps in single/multiple selection from list of data.
+///
+/// {@template arguments}
+/// The [listData] should be list of dynamic data which neeeds to filter.
+///
+/// The [selectedListData] should be subset of [listData]. The list passed to [selectedListData] should available in [listData].
+///
+/// The [choiceChipLabel] is a callback which required [String] value in return. It used this value to display text on choice chip.
+///
+/// The [validateSelectedItem] used to identifies weather a item is selecte or not.
+///
+/// [onItemSearch] filter the list on the basis of search field text. It expose search api to permform search operation accoding to requirement.
+/// When text change in search text field then return a list of element which contains specific text. if no element found then it should return empty list.
+///
+/// ```dart
+///    onItemSearch: (list, text) {
+///     if (list.any((element) =>
+///         element.toLowerCase().contains(text.toLowerCase()))) {
+///       /// return list which contains matches
+///       return list
+///           .where((element) =>
+///               element.toLowerCase().contains(text.toLowerCase()))
+///           .toList();
+///     }
+///   },
+/// ```
+///
+/// The [choiceChipBuilder] is a builder to design custom choice chip.
+///
+///
+/// The [onApplyButtonClick] is a callback which return list of all selected items on apply button click.  if no item is selected then it will return empty list.
+/// {@endtemplate}
+/// The [useSafeArea] argument is used to indicate if the dialog should only display in 'safe' areas of the screen not used by the operating system (see [SafeArea] for more details). It is true by default, which means the dialog will not overlap operating system areas. If it is set to false the dialog will only be constrained by the screen size. It can not be null.
+///
+/// The [useRootNavigator] argument is used to determine whether to push the dialog to the [Navigator] furthest from or nearest to the given context. By default, useRootNavigator is true and the dialog route created by this method is pushed to the root navigator. It can not be null.
+///
+///The [routeSettings] argument is passed to [showGeneralDialog], see [RouteSettings] for details.
+///
+/// ### This example shows how to use [FilterListDialog]
+///
+///  ``` dart
+/// void _openFilterDialog() async {
+///    await FilterListDialog.display<String>(context,
+///        listData: ["One", "Two", "Three", "Four","five","Six","Seven","Eight","Nine","Ten"],
+///        selectedListData: ["One", "Three", "Four","Eight","Nine"],
+///        choiceChipLabel: (item) {
+///          return item;
+///        },
+///        validateSelectedItem: (list, val) {
+///          return list!.contains(val);
+///        },
+///        onItemSearch: (list, text) {
+///          if (list!.any((element) =>
+///              element.toLowerCase().contains(text.toLowerCase()))) {
+///            /// return list which contains text matches
+///            return list
+///                .where((element) =>
+///                    element.toLowerCase().contains(text.toLowerCase()))
+///                .toList();
+///          }
+///          return [];
+///        },
+///        height: 480,
+///        borderRadius: 20,
+///        headlineText: "Select Count",
+///        searchFieldHintText: "Search Here",
+///        onApplyButtonClick: (list) {
+///          if (list != null) {
+///            setState(() {
+///             var selectedList = List.from(list);
+///            });
+///            Navigator.pop(context);
+///          }
+///        });
+///  }
+/// ```
+
 class FilterListDialog {
-  /// {@tool snippet}
-  ///
-  /// This example shows how to use [FilterListDialog]
-  ///
-  ///  ``` dart
-  /// void _openFilterDialog() async {
-  ///   await FilterListDialog.display(context,
-  ///       listData: ["One", "Two", "Three", "Four","five","Six","Seven","Eight","Nine","Ten"],
-  ///       selectedListData: ["One", "Three", "Four","Eight","Nine"],
-  ///       label: (item) {
-  ///         return item;
-  ///       },
-  ///       validateSelectedItem: (list, val) {
-  ///         return list.contains(val);
-  ///       },
-  ///       onItemSearch: (list, text) {
-  ///         if (list.any((element) =>
-  ///             element.toLowerCase().contains(text.toLowerCase()))) {
-  ///           /// return list which contains matches
-  ///           return list
-  ///               .where((element) =>
-  ///                   element.toLowerCase().contains(text.toLowerCase()))
-  ///               .toList();
-  ///         }
-  ///       },
-  ///       height: 480,
-  ///       borderRadius: 20,
-  ///       headlineText: "Select Count",
-  ///       searchFieldHintText: "Search Here",
-  ///       onApplyButtonClick: (list) {
-  ///         if (list != null) {
-  ///           setState(() {
-  ///            var selectedList = List.from(list);
-  ///           });
-  ///           Navigator.pop(context);
-  ///         }
-  ///       });
-  /// }
-  /// ```
-  /// {@end-tool}
-  ///
   static Future display<T>(context,
       {
 
-      /// Pass list containing all data which neeeds to filter
-      @required List<T> listData,
+      /// Pass list containing all data which neeeds to filter.
+      required List<T> listData,
 
       /// pass selected list of object
-      /// every object on selecteListData should be present in list data
-      List<T> selectedListData,
+      /// every object on selecteListData should be present in list data.
+      List<T>? selectedListData,
 
-      /// Display text value on choice chip
-      @required String Function(T b) label,
+      /// Display text on choice chip.
+      required LabelDelegate<T> choiceChipLabel,
 
-      /// identifies weather a item is selecte or not
-      @required ValidateSelectedItem<T> validateSelectedItem,
+      /// identifies weather a item is selecte or not.
+      required ValidateSelectedItem<T> validateSelectedItem,
 
-      /// filter list on the basis of search field text
-      /// When text change in search text field then return list containing that text value
+      /// filter list on the basis of search field text.
+      /// When text change in search text field then return list containing that text value.
       ///
-      ///Check if list has value which matches to text
-      @required List<T> Function(List<T> list, String text) onItemSearch,
+      ///Check if list has value which matches to text.
+      required ItemSearchDelegate<T> onItemSearch,
 
       /// Return list of all selected items
-      @required OnApplyButtonClick<T> onApplyButtonClick,
-      double height,
-      double width,
+      required OnApplyButtonClick<T> onApplyButtonClick,
+
+      /// Height of the dialog
+      double? height,
+
+      /// Width of the dialog
+      double? width,
+
+      /// Border radius of dialog.
       double borderRadius = 20,
+
+      /// Headline text to be display as header of dialog.
       String headlineText = "Select",
+
+      /// Hint text for search field.
       String searchFieldHintText = "Search here",
+
+      /// Used to hide selected text count.
       bool hideSelectedTextCount = false,
+
+      /// Used to hide search field.
       bool hideSearchField = false,
-      bool hidecloseIcon = false,
+
+      /// Used to hide close icon.
+      bool hideCloseIcon = false,
+
+      /// Used to hide header.
       bool hideheader = false,
-      bool hideheaderText = false,
+
+      /// Used to hide header text.
+      bool hideHeaderText = false,
+
+      ///Color of close icon
       Color closeIconColor = Colors.black,
+
+      /// The `barrierDismissible` argument is used to indicate whether tapping on the barrier will dismiss the dialog.
+      ///
+      ///  It is true by default and can not be null.
       bool barrierDismissible = true,
       bool useSafeArea = true,
       bool useRootNavigator = true,
-      RouteSettings routeSettings,
+      RouteSettings? routeSettings,
 
-      /// if `enableOnlySingleSelection` is true then it disabled the multiple selection
+      /// if `enableOnlySingleSelection` is true then it disabled the multiple selection.
       /// and enabled the single selection model.
       ///
-      /// Defautl value is `false`
+      /// Defautl value is [false]
       bool enableOnlySingleSelection = false,
 
-      /// Background color of dialog box
+      /// Background color of dialog box.
       Color backgroundColor = Colors.white,
 
-      /// Background color for search field
+      /// Background color for search field.
       Color searchFieldBackgroundColor = const Color(0xfff5f5f5),
 
-      /// Background color for Apply button
+      /// Background color for Apply button.
       Color applyButonTextBackgroundColor = Colors.blue,
 
-      /// TextStyle for chip when selected
-      TextStyle selectedChipTextStyle,
+      /// TextStyle for chip when selected.
+      TextStyle? selectedChipTextStyle,
 
-      /// TextStyle for chip when not selected
-      TextStyle unselectedChipTextStyle,
+      /// TextStyle for chip when not selected.
+      TextStyle? unselectedChipTextStyle,
 
-      /// TextStyle for `All` and `Reset` button text
-      TextStyle controlButtonTextStyle,
+      /// TextStyle for [All] and [Reset] button text.
+      TextStyle? controlButtonTextStyle,
 
-      /// TextStyle for `Apply` button
-      TextStyle applyButtonTextStyle,
+      /// TextStyle for [Apply] button.
+      TextStyle? applyButtonTextStyle,
 
-      /// TextStyle for header text
-      TextStyle headerTextStyle,
+      /// TextStyle for header text.
+      TextStyle? headerTextStyle,
 
-      /// TextStyle for search field text
-      TextStyle searchFieldTextStyle,
+      /// TextStyle for search field text.
+      TextStyle? searchFieldTextStyle,
 
-      /// Builder for custom choice chip
-      ChoiceChipBuilder choiceChipBuilder}) async {
-    assert(validateSelectedItem != null, ''' 
-            validateSelectedItem callback can not be null
-
-            Tried to use below callback to ignore error.
-
-             validateSelectedItem: (list, val) {
-                  return list.contains(val);
-             }
-            ''');
+      /// The `choiceChipBuilder` is a builder to design custom choice chip.
+      ChoiceChipBuilder? choiceChipBuilder}) async {
     if (height == null) {
       height = MediaQuery.of(context).size.height * .8;
     }
@@ -155,7 +208,7 @@ class FilterListDialog {
             color: Colors.transparent,
             child: FilterListWidget(
               listData: listData,
-              label: label,
+              choiceChipLabel: choiceChipLabel,
               width: width,
               height: height,
               hideHeader: hideheader,
@@ -169,8 +222,8 @@ class FilterListDialog {
               onApplyButtonClick: onApplyButtonClick,
               validateSelectedItem: validateSelectedItem,
               hideSelectedTextCount: hideSelectedTextCount,
-              hidecloseIcon: hidecloseIcon,
-              hideheaderText: hideheaderText,
+              hideCloseIcon: hideCloseIcon,
+              hideHeaderText: hideHeaderText,
               hideSearchField: hideSearchField,
               choiceChipBuilder: choiceChipBuilder,
               searchFieldHintText: searchFieldHintText,