+++ lib/src/widget/choice_chip_widget.dart
@@ -3,7 +3,7 @@ import 'package:flutter/material.dart';
 
 class ChoiceChipWidget<T> extends StatelessWidget {
   const ChoiceChipWidget(
-      {Key key,
+      {Key? key,
       this.text,
       this.item,
       this.selected,
@@ -15,20 +15,20 @@ class ChoiceChipWidget<T> extends StatelessWidget {
       this.unselectedChipTextStyle})
       : super(key: key);
 
-  final String text;
-  final bool selected;
-  final Function(bool) onSelected;
-  final Color unselectedTextBackgroundColor;
-  final Color selectedTextBackgroundColor;
-  final TextStyle selectedChipTextStyle;
-  final TextStyle unselectedChipTextStyle;
-  final T item;
+  final String? text;
+  final bool? selected;
+  final Function(bool)? onSelected;
+  final Color? unselectedTextBackgroundColor;
+  final Color? selectedTextBackgroundColor;
+  final TextStyle? selectedChipTextStyle;
+  final TextStyle? unselectedChipTextStyle;
+  final T? item;
 
   /// Builder for custom choice chip
-  final ChoiceChipBuilder choiceChipBuilder;
+  final ChoiceChipBuilder? choiceChipBuilder;
 
   TextStyle getSelectedTextStyle(BuildContext context) {
-    return selected
+    return selected!
         ? selectedChipTextStyle ??
             TextStyle(color: Theme.of(context).colorScheme.onPrimary)
         : unselectedChipTextStyle ?? TextStyle(color: Colors.black);
@@ -39,24 +39,24 @@ class ChoiceChipWidget<T> extends StatelessWidget {
     return choiceChipBuilder != null
         ? GestureDetector(
             onTap: () {
-              onSelected(true);
+              onSelected!(true);
             },
-            child: choiceChipBuilder(context, item, selected),
+            child: choiceChipBuilder!(context, item, selected),
           )
         : Container(
             padding: const EdgeInsets.symmetric(horizontal: 10),
             child: ChoiceChip(
-              backgroundColor: selected
+              backgroundColor: selected!
                   ? selectedTextBackgroundColor
                   : unselectedTextBackgroundColor,
-              selectedColor: selected
+              selectedColor: selected!
                   ? selectedTextBackgroundColor
                   : unselectedTextBackgroundColor,
               label: Text(
                 '$text',
                 style: getSelectedTextStyle(context),
               ),
-              selected: selected,
+              selected: selected!,
               onSelected: onSelected,
             ),
           );